arp and lb unit tests
