+++ src/passthrough/lb.rs
@@ -13,9 +13,10 @@ use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 use std::sync::{Arc, Mutex, RwLock};
 use std::sync::mpsc::Sender;
 use std::collections::HashMap;
+use std::str::FromStr;
 use lru_time_cache::LruCache;
 use self::passthrough::utils::{IPV4_HEADER_LEN, EPHEMERAL_PORT_LOWER, EPHEMERAL_PORT_UPPER};
-
+use crate::config::{Config};
 
 #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Debug)]
 pub struct Client {
@@ -79,8 +80,79 @@ pub struct LB {
     pub stats_update_frequency: u64,
 }
 
-
 impl LB {
+    pub fn new(frontend_name: String, conf: Config, dsr: bool) -> Option<LB> {
+        let front = conf.base.frontends.get(&frontend_name).unwrap();
+
+        let mut backend_servers = HashMap::new();
+
+        // Set defaults
+        let mut health_check_interval = 5;
+        let mut connection_tracker_capacity = 1000 as usize;
+        let mut workers = 4 as usize;
+        let mut stats_update_frequency = 5;
+
+        match conf.base.passthrough {
+            Some(setting) => {
+                connection_tracker_capacity = setting.connection_tracker_capacity;
+                if let Some(num) = setting.workers {
+                    workers = num;
+                }
+                if let Some(freq) = setting.stats_update_frequency {
+                    stats_update_frequency = freq;
+                }
+            },
+            None => {},
+        }
+
+        match conf.base.backends.get(&front.backend) {
+            Some(back) => {
+                for (_,addr) in &back.servers {
+                    let listen_addr: SocketAddr = FromStr::from_str(&addr.addr)
+                                      .ok()
+                                      .expect("Failed to parse listen host:port string");
+                    backend_servers.insert(listen_addr, addr.weight);
+                }
+                if back.health_check_interval > 0 {
+                    health_check_interval = back.health_check_interval;
+                }
+            }
+            None => error!("Error finding backend server pool in config: {} not found on backend config", front.backend),
+        };
+
+        if backend_servers.len() > 0 {
+            let listen_addr: SocketAddr = FromStr::from_str(&front.listen_addr)
+                              .ok()
+                              .expect("Failed to parse listen host:port string");
+
+            let backend = Arc::new(Backend::new(front.backend.clone(), backend_servers, health_check_interval));
+            match listen_addr.ip() {
+                IpAddr::V4(ip4) => {
+                    let new_lb = LB {
+                        name: frontend_name.clone(),
+                        listen_ip: ip4,
+                        listen_port: listen_addr.port(),
+                        backend: backend.clone(),
+                        conn_tracker: Arc::new(RwLock::new(LruCache::<Client, Connection>::with_capacity(connection_tracker_capacity))),
+                        port_mapper: Arc::new(RwLock::new(HashMap::new())),
+                        next_port: Arc::new(Mutex::new(EPHEMERAL_PORT_LOWER)),
+                        workers: workers,
+                        dsr: dsr,
+                        stats_update_frequency: stats_update_frequency,
+                    };
+                    return Some(new_lb)
+                }
+                _ => {
+                    error!("Unable to configure load balancer server {:?}.  Only Ipv4 is supported", front);
+                    return None
+                }
+            }
+        } else {
+            error!("Unable to configure load balancer server {:?}", front);
+            return None
+        }
+    }
+
     pub fn next_avail_port(&mut self) -> u16 {
         let mut port = self.next_port.lock().unwrap();
         if *port < EPHEMERAL_PORT_UPPER {
@@ -294,5 +366,148 @@ impl LB {
             return Some(conn.clone());
         }
         None
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    extern crate hyper;
+    use std::sync::mpsc::channel;
+    use crate::config::{Config};
+    use crate::passthrough;
+    use std::net::{IpAddr, Ipv4Addr, SocketAddr};
+    use self::passthrough::utils::{EPHEMERAL_PORT_LOWER, EPHEMERAL_PORT_UPPER, build_dummy_ip};
+    use self::passthrough::backend::Node;
+    use pnet::packet::tcp::{TcpPacket, MutableTcpPacket};
+    use pnet::packet::ipv4::{MutableIpv4Packet};
+    use pnet::packet::Packet;
+
+    #[test]
+    fn test_new_lb() {
+        let conf = Config::new("testdata/passthrough_test.toml").unwrap();
+        let test_lb = passthrough::LB::new("tcp_3000".to_string(), conf, false).unwrap();
+
+        let ip: Ipv4Addr = "127.0.0.1".parse().unwrap();
+        assert_eq!(test_lb.listen_ip, ip);
+        assert_eq!(test_lb.listen_port, 3000);
+        assert_eq!(test_lb.name, "tcp_3000");
+    }
+
+    #[test]
+    fn test_next_port() {
+        let conf = Config::new("testdata/passthrough_test.toml").unwrap();
+        let mut test_lb = passthrough::LB::new("tcp_3000".to_string(), conf, false).unwrap();
+
+        let first_port = test_lb.next_avail_port();
+        assert_eq!(*test_lb.next_port.lock().unwrap(), first_port);
+        assert_eq!(test_lb.next_avail_port(), first_port + 1);
+        *test_lb.next_port.lock().unwrap() = EPHEMERAL_PORT_UPPER + 1;
+        assert_eq!(test_lb.next_avail_port(), EPHEMERAL_PORT_LOWER);
+    }
+
+    #[test]
+    fn test_next_client_handler() {
+        let conf = Config::new("testdata/passthrough_test.toml").unwrap();
+        let mut test_lb = passthrough::LB::new("tcp_3000".to_string(), conf, false).unwrap();
+
+        let (tx, rx) = channel();
+        let lb_ip = "127.0.0.1".parse().unwrap();
+        let client_ip: Ipv4Addr = "9.9.9.9".parse().unwrap();
+        let backend_srv_ip: Ipv4Addr = "127.0.0.1".parse().unwrap();
+
+        {
+            // set a backend server to healthy
+            let mut srvs_map = test_lb.backend.servers_map.write().unwrap();
+            let mut srvs_ring = test_lb.backend.ring.lock().unwrap();
+            let health = srvs_map.get_mut(&SocketAddr::new(IpAddr::V4(backend_srv_ip), 3080)).unwrap();
+            *health = true;
+            srvs_ring.add_node(&Node{host: IpAddr::V4(backend_srv_ip), port: 3080})
+        }
+
+        // gen test ip/tcp packet with simulated client
+        let req_header = build_dummy_ip(client_ip, lb_ip, 43000, 3000);
+        let mut tcp_header = MutableTcpPacket::owned(req_header.payload().to_owned()).unwrap();
+
+        // call client_handler and verify packet being sent out to healthy backend server
+        test_lb.client_handler(&mut req_header.to_immutable(), &mut tcp_header, tx.clone());
+        let fwd_pkt: MutableIpv4Packet = rx.recv().unwrap();
+        assert_eq!(fwd_pkt.get_destination(), backend_srv_ip);
+        assert_eq!(fwd_pkt.get_source(), lb_ip);
+
+        let tcp_resp = TcpPacket::new(fwd_pkt.payload()).unwrap();
+        assert_eq!(tcp_resp.get_destination(), 3080);
+        assert_eq!(tcp_resp.get_source(), EPHEMERAL_PORT_LOWER + 1);
+
+        {
+            // check connection is being tracked
+            let port_mp = test_lb.port_mapper.read().unwrap();
+            let cli = port_mp.get(&(EPHEMERAL_PORT_LOWER + 1)).unwrap();
+
+            let mut test_lb = test_lb.conn_tracker.write().unwrap();
+            let conn = test_lb.get(&cli).unwrap();
+            assert_eq!(conn.ephem_port, EPHEMERAL_PORT_LOWER + 1);
+            assert_eq!(conn.client, SocketAddr::new(IpAddr::V4(client_ip), 43000));
+        }
+
+        {
+            assert_eq!(test_lb.conn_tracker.read().unwrap().len(), 1);
+        }
+
+        {
+            // check same client again to verify connection tracker is used
+            let mut tcp_header = MutableTcpPacket::owned(req_header.payload().to_owned()).unwrap();
+            test_lb.client_handler(&mut req_header.to_immutable(), &mut tcp_header, tx.clone());
+            // next port should not have incremented
+            assert_eq!(*test_lb.next_port.lock().unwrap(), EPHEMERAL_PORT_LOWER + 1);
+
+            let fwd_pkt: MutableIpv4Packet = rx.recv().unwrap();
+            assert_eq!(fwd_pkt.get_destination(), backend_srv_ip);
+            assert_eq!(fwd_pkt.get_source(), lb_ip);
+
+            let tcp_resp = TcpPacket::new(fwd_pkt.payload()).unwrap();
+            assert_eq!(tcp_resp.get_destination(), 3080);
+            assert_eq!(tcp_resp.get_source(), EPHEMERAL_PORT_LOWER + 1);
+            assert_eq!(test_lb.conn_tracker.read().unwrap().len(), 1);
+        }
+
+        {
+            // set backend server to unhealthy
+            let mut srvs_map = test_lb.backend.servers_map.write().unwrap();
+            let mut srvs_ring = test_lb.backend.ring.lock().unwrap();
+            let health = srvs_map.get_mut(&SocketAddr::new(IpAddr::V4(backend_srv_ip), 3080)).unwrap();
+            *health = false;
+            srvs_ring.remove_node(&Node{host: IpAddr::V4(backend_srv_ip), port: 3080})
+        }
+
+        // check same client again to verify connection is failed
+        let mut tcp_header = MutableTcpPacket::owned(req_header.payload().to_owned()).unwrap();
+        test_lb.client_handler(&mut req_header.to_immutable(), &mut tcp_header, tx.clone());
+        // since there are not healthy backend servers there should be no connections added to map
+        assert_eq!(test_lb.conn_tracker.read().unwrap().len(), 0);
+    }
+
+    #[test]
+    fn test_passthrough_server_response() {
+        let conf = Config::new("testdata/passthrough_test.toml").unwrap();
+        let srv = passthrough::Server::new(conf, false);
+        let mut lb = srv.lbs[0].clone();
+
+        let (tx, rx) = channel();
+        let lb_ip = "127.0.0.1".parse().unwrap();
+        let client_ip: Ipv4Addr = "9.9.9.9".parse().unwrap();
+        let backend_srv_ip: Ipv4Addr = "8.8.8.8".parse().unwrap();
+
+        // simulate response from server at port 80 to local (ephemeral) port 35000
+        let resp_header = build_dummy_ip(backend_srv_ip, lb_ip, 80, 35000);
+        let mut tcp_header = MutableTcpPacket::owned(resp_header.payload().to_owned()).unwrap();
+        // server should respond to client ip at client's port
+        lb.server_response_handler(&resp_header.to_immutable(), &mut tcp_header, &SocketAddr::new(IpAddr::V4(client_ip), 55000), tx);
+        let srv_resp: MutableIpv4Packet = rx.recv().unwrap();
+        assert_eq!(srv_resp.get_destination(), client_ip);
+        assert_eq!(srv_resp.get_source(), lb_ip);
+
+        let tcp_resp = TcpPacket::new(srv_resp.payload()).unwrap();
+        assert_eq!(tcp_resp.get_destination(), 55000);
+        assert_eq!(tcp_resp.get_source(), 3000);
     }
 }
