+++ src/passthrough/utils.rs
@@ -19,7 +19,7 @@ pub const ETHERNET_HEADER_LEN: usize = 14;
 pub const IPV4_HEADER_LEN: usize = 20;
 pub const TCP_HEADER_LEN: usize = 32;
 
-fn find_local_addr() -> Option<IpAddr> {
+pub fn find_local_addr() -> Option<IpAddr> {
     for iface in pnet::datalink::interfaces() {
         if !iface.is_loopback() {
             for ipnet in iface.ips {
